Refactor user functions to lib package
